remove remaining AggregateStore support code
